fixed so it only renames txt files

+++ FileRenamer.py
@@ -1,4 +1,4 @@
-import os
+﻿import os
 
 folder_path = r"C:\Users\busch\OneDrive\Documents\aagergaergaerg\PYYYTest"
 
@@ -6,11 +6,12 @@ files = os.listdir(folder_path)
 
 for index, file in enumerate(files):
     old_path = os.path.join(folder_path, file)
-    new_name = f"file_{index+1}.txt"
-    new_path = os.path.join(folder_path, new_name)
 
-    os.rename(old_path, new_path)
-    print(f"Renamed: {file} -> {new_name}")
+    if os.path.isfile(old_path):
+        new_name = f"file_{index+1}.txt"
+        new_path = os.path.join(folder_path, new_name)
 
+        os.rename(old_path, new_path)
+        print(f"Renamed: {file} -> {new_name}")
 
-print("All files renamed successfully !")
+print("All files renamed successfully (folders ignored)!")
